requirements.txt
fastapi
uvicorn
pydantic
pandas
openpyxl
selenium
beautifulsoup4
webdriver_manager
python-dotenv